+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 _Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v5.x.x_
 
+BUG FIXES:
+
+- Fix retry transact data on redis doesn't get removed after successful retry.
+
 OTHERS:
 
 - [Main] Add environment variable option alias `TELEMETRY_DB_HOST` for `TELEMETRY_DB_IP` to match with telemetry client module.